Fixes issue #41
Adds a override to EquipmentAsset's search method introduced from
acts_as_searchable. Properly handles visibility based on scope of search
and permissions.

The search algorithm assumed that all searchable models are associated
with a project which allowed it to use project based scope on a search.
Since EquipmentAsset is not associated with a project we had to exclude
searching EquipmentAssets when the user is doing a project search and
not a global search as one would if searching from the index page.